enhance: auto layout hasSider (#45361)
* enhance: auto layout hasSider

* test: update snapshot
