+++ components/layout/Sider.tsx
@@ -1,14 +1,14 @@
+import * as React from 'react';
+import { useContext, useEffect, useRef, useState } from 'react';
 import BarsOutlined from '@ant-design/icons/BarsOutlined';
 import LeftOutlined from '@ant-design/icons/LeftOutlined';
 import RightOutlined from '@ant-design/icons/RightOutlined';
 import classNames from 'classnames';
 import omit from 'rc-util/lib/omit';
-import * as React from 'react';
-import { useContext, useEffect, useRef, useState } from 'react';
 
 import isNumeric from '../_util/isNumeric';
 import { ConfigContext } from '../config-provider';
-import { LayoutContext } from './layout';
+import { LayoutContext } from './context';
 
 const dimensionMaxMap = {
   xs: '479.98px',
