+++ components/layout/context.ts
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+export interface LayoutContextProps {
+  siderHook: {
+    addSider: (id: string) => void;
+    removeSider: (id: string) => void;
+  };
+}
+
+export const LayoutContext = React.createContext<LayoutContextProps>({
+  siderHook: {
+    addSider: () => null,
+    removeSider: () => null,
+  },
+});
